rename io_xml to io_fews and reorganise fews functions
